requirements.txt
python-telegram-bot
nest_asyncio
aiosqlite
aiohttp
